Begin refactoring into different mesh classes
